feat(homepage): fold "View as" into Preview mode as inline switcher
Preview mode now carries a "Viewing as" segmented control (Everyone /
User / Group / Role). Everyone shows the live draft; picking a
user/group/role resolves the homepage they actually land on through the
real precedence chain and renders it (or the day-one default / dashboard
redirect). Removes the separate "View as" toolbar button and modal.
